+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   Added `gatewayId` to `FormSubmissionEventConfiguration`
+
 ## 1.6.0 (2020-06-10)
 
 ### Added